### 2024-09-07 #7192 @editorialbot
[PRE REVIEW]: AdditiveFOAM 1.0: A Continuum Physics Code for Additive Manufacturing
**Submitting author:** @colemanjs (<a href="http://orcid.org/0000-0002-7261-3143">John Coleman</a>)
**Repository:** https://github.com/ORNL/AdditiveFOAM.git
**Branch with paper.md** (empty if default branch): joss-2024
**Version:** v1.0
