### 2024-10-04 8c875cd
Update _config.yml
attempt to remove old customization
