+++ _config.yml
@@ -57,6 +57,6 @@ callouts:
     color: ornl_green
 
 # Update custom style assets
-assets:
-  stylesheets:
-    - /assets/css/custom.css
+#assets:
+#  stylesheets:
+#    - /assets/css/custom.css
