### 2024-10-04 17ad39b
Update _config.yml to update custom style assets
