+++ _config.yml
@@ -57,6 +57,8 @@ callouts:
     color: ornl_green
 
 # Update custom style assets
+sass:
+  custom_stylesheet: _sass/custom/custom.scss
 #assets:
 #  stylesheets:
 #    - /assets/css/custom.css
