+++ _config.yml
@@ -55,10 +55,3 @@ callouts:
   custom:
     title: Note
     color: ornl_green
-
-# Update custom style assets
-sass:
-  custom_stylesheet: _sass/custom/custom.scss
-#assets:
-#  stylesheets:
-#    - /assets/css/custom.css
