### 2024-10-04 ece8ba5
Update _config.yml to include custom.scss
