### 2024-10-04 9869de2
Update README.md order to prioritize documentation
