### 2024-10-04 653aae6
reorganized user guide layout
