+++ docs/theory.md
@@ -1,6 +1,7 @@
 ---
+parent: AdditiveFOAM v1.0 User Guide
 title: Theory
-nav_order: 3
+nav_order: 2
 usemathjax: true
 ---
 
