+++ docs/usage.md
@@ -1,6 +1,7 @@
 ---
+parent: AdditiveFOAM v1.0 User Guide
 title: Usage
-nav_order: 5
+nav_order: 3
 usemathjax: true
 ---
 
