+++ docs/index.md
@@ -1,14 +1,13 @@
 ---
-title: AdditiveFOAM User Guide
+title: AdditiveFOAM v1.0 User Guide
 nav_order: 2
+has_children: true
 ---
 
-# AdditiveFOAM User Guide
+# AdditiveFOAM v1.0 User Guide
 AdditiveFOAM is a free, open source heat and mass transfer software for simulations of Additive Manufacturing (AM) released by Oak Ridge National Laboratory. It is built upon [OpenFOAM](https://openfoam.org/), a free, open source computational fluid dynamics (CFD) software package released by the OpenFOAM Foundation.
 
 ---
 
-## Contents
-- [Theory](theory.md)
-- [Installation](installation.md)
-- [Usage](usage.md)
+{: .custom }
+The User Guide v1.0 is currently under development and may be incomplete.
