+++ _config.yml
@@ -7,7 +7,8 @@ repository: ORNL/AdditiveFOAM
 
 permalink: pretty
 
-remote_theme: just-the-docs/just-the-docs
+#remote_theme: just-the-docs/just-the-docs
+theme: just-the-docs
 
 plugins:
   - jekyll-remote-theme
