+++ _config.yml
@@ -7,8 +7,7 @@ repository: ORNL/AdditiveFOAM
 
 permalink: pretty
 
-#remote_theme: just-the-docs/just-the-docs
-theme: just-the-docs
+remote_theme: just-the-docs/just-the-docs
 
 plugins:
   - jekyll-remote-theme
