### 2024-10-04 4383a88
Add table of contents to user guide
