### 2024-10-30 f5044e8
corrected explicit energy solve forced to remain forward euler. Also, added ExaCA support for all heat transfer tutorials
