### 2024-11-06 a447598
made foamToExaCA a functionObject ExaCA to remain consistent with OpenFOAM library structure
