+++ applications/solvers/additiveFoam/additiveFoam.C
@@ -40,7 +40,6 @@ Description
 #include "interpolateXY/interpolateXY.H"
 
 #include "movingHeatSourceModel.H"
-#include "foamToExaCA/foamToExaCA.H"
 
 #include "EulerDdtScheme.H"
 #include "CrankNicolsonDdtScheme.H"
@@ -65,8 +64,6 @@ int main(int argc, char *argv[])
     scalar DiNum = 0.0;
 
     scalar alphaCoNum = 0.0;
-
-    foamToExaCA ExaCA(T);
 
     movingHeatSourceModel sources(mesh);
 
@@ -98,16 +95,12 @@ int main(int argc, char *argv[])
 
         #include "thermo/TEqn.H"
         
-        ExaCA.update();
-
         runTime.write();
 
         Info<< "ExecutionTime = " << runTime.elapsedCpuTime() << " s"
             << "  ClockTime = " << runTime.elapsedClockTime() << " s"
             << nl << endl;
     }
-
-    ExaCA.write();
 
     return 0;
 }
