### 2024-11-06 b07f5ad
updated build instructions for CI
