### 2024-11-07 f484321
corrected explicit energy solve forced to remain forward euler. (#37)
* corrected explicit energy solve forced to remain forward euler. Also, added ExaCA support for all heat transfer tutorials

* made foamToExaCA a functionObject ExaCA to remain consistent with OpenFOAM library structure

* updated build instructions for CI
