+++ applications/solvers/additiveFoam/functionObjects/ExaCA/ExaCA.C
@@ -1,0 +1,463 @@
+/*---------------------------------------------------------------------------*\
+  =========                 |
+  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
+   \\    /   O peration     | Website:  https://openfoam.org
+    \\  /    A nd           | Copyright (C) 2024 OpenFOAM Foundation
+     \\/     M anipulation  |
+-------------------------------------------------------------------------------
+                Copyright (C) 2023 Oak Ridge National Laboratory                
+-------------------------------------------------------------------------------
+License
+    This file is part of OpenFOAM.
+
+    OpenFOAM is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    OpenFOAM is distributed in the hope that it will be useful, but WITHOUT
+    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+    for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
+
+\*---------------------------------------------------------------------------*/
+
+#include "ExaCA.H"
+#include "Time.H"
+#include "fvMesh.H"
+#include "addToRunTimeSelectionTable.H"
+#include "volFields.H"
+#include "OFstream.H"
+#include "OSspecific.H"
+#include "meshSearch.H"
+#include "labelVector.H"
+#include "pointMVCWeight.H"
+
+// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
+
+namespace Foam
+{
+namespace functionObjects
+{
+    defineTypeNameAndDebug(ExaCA, 0);
+    
+    addToRunTimeSelectionTable
+    (
+        functionObject,
+        ExaCA,
+        dictionary
+    );
+}
+}
+
+// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
+
+Foam::functionObjects::ExaCA::ExaCA
+(
+    const word& name,
+    const Time& runTime,
+    const dictionary& dict
+)
+:
+    fvMeshFunctionObject(name, runTime, dict),
+    T_(mesh_.lookupObject<VolField<scalar>>("T")),
+    vpi_(mesh_),
+    Tp_
+    (
+        IOobject
+        (
+            "Tp_",
+            runTime.timeName(),
+            mesh_,
+            IOobject::NO_READ,
+            IOobject::NO_WRITE
+        ),
+        vpi_.interpolate(T_)
+    )
+{
+    read(dict);
+    
+    setOverlapCells();
+}
+
+
+// * * * * * * * * * * * * * * * * Destructor  * * * * * * * * * * * * * * * //
+
+Foam::functionObjects::ExaCA::~ExaCA()
+{}
+
+
+// * * * * * * * * * * * * * * * Member Functions  * * * * * * * * * * * * * //
+
+bool Foam::functionObjects::ExaCA::read(const dictionary& dict)
+{
+    box_ = dict.lookup("box");
+    
+    isoValue_ = dict.lookup<scalar>("isoValue");
+    
+    dx_  = dict.lookup<scalar>("dx");
+
+    return true;
+}
+
+void Foam::functionObjects::ExaCA::setOverlapCells()
+{
+    // create a compact cell-stencil using the overlap sub-space
+    const pointField& points = mesh_.points();
+
+    boundBox procBb(points);
+
+    const vector extend = 1e-10*vector::one;
+
+    if (procBb.overlaps(box_))
+    {
+        forAll(mesh_.cells(), celli)
+        {
+            boundBox cellBb(point::max, point::min);
+
[…]
