+++ applications/solvers/additiveFoam/functionObjects/ExaCA/ExaCA.H
@@ -2,7 +2,7 @@
   =========                 |
   \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
    \\    /   O peration     | Website:  https://openfoam.org
-    \\  /    A nd           | Copyright (C) 2011-2022 OpenFOAM Foundation
+    \\  /    A nd           | Copyright (C) 2024 OpenFOAM Foundation
      \\/     M anipulation  |
 -------------------------------------------------------------------------------
                 Copyright (C) 2023 Oak Ridge National Laboratory                
@@ -24,44 +24,35 @@ License
     along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
 
 Class
-    Foam::foamToExaca
-
-Description
-    A utility for interpolating temperature data to a uniform cartestian grid
-    and writes the data in the ExaCA data format.
+    Foam::functionObjects::ExaCA
 
 SourceFiles
-    foamToExaca.C
+    ExaCA.C
 
 \*---------------------------------------------------------------------------*/
 
-#ifndef foamToExaCA_H
-#define foamToExaCA_H
-    
-#include "IOdictionary.H"
-#include "volFields.H"
+#ifndef ExaCA_H
+#define ExaCA_H
+
+#include "fvMeshFunctionObject.H"
 #include "volPointInterpolation.H"
-#include "meshSearch.H"
 
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 namespace Foam
 {
+namespace functionObjects
+{
 
 /*---------------------------------------------------------------------------*\
-         Class foamToExaca Declaration
+                       Class ExaCA Declaration
 \*---------------------------------------------------------------------------*/
 
-class foamToExaCA
+class ExaCA
 :
-    public IOdictionary
+    public fvMeshFunctionObject
 {
-private:
-    // Private data
-
-        const fvMesh& mesh_;
-
-        const Time& runTime_;
+    // Private Data
 
         const volScalarField& T_;
 
@@ -69,60 +60,88 @@ private:
 
         mutable pointScalarField Tp_;
 
+        DynamicList<label> overlapCells;
+
+        DynamicList<List<scalar>> events;
 
         List<DynamicList<label>> pointsInCell;
 
         DynamicList<point> positions;
+        
         DynamicList<scalarField> weights;
-
-        DynamicList<label> compactCells;
-
-        DynamicList<List<scalar>> events;
-
-        Switch    execute_;
-
+              
         boundBox  box_;
 
-        scalar    iso_;
+        scalar    isoValue_;
 
         scalar    dx_;
 
-public:
-    TypeName("foamToExaCA");
+    // Private Member Functions
 
-    // Construct from temperature field
-    foamToExaCA(const volScalarField& T);
+
+public:
+
+    //- Runtime type information
+    TypeName("ExaCA");
+
+
+    // Constructors
+
+        //- Construct from Time and dictionary
+        ExaCA
+        (
+            const word& name,
+            const Time& runTime,
+            const dictionary& dict
+        );
+
+        //- Disallow default bitwise copy construction
+        ExaCA(const ExaCA&) = delete;
+
 
     //- Destructor
-    virtual ~foamToExaCA()
-    {}
+    virtual ~ExaCA();
+
 
     // Member Functions
-    virtual void initialize();
 
-    virtual void update();
+        //- Read the ExaCA data
+        virtual bool read(const dictionary&);
 
-    virtual void write();
+        //- Initialize lists
+        virtual void setOverlapCells();//initialize();
 
-    virtual void mapPoints(const meshSearch& searchEngine);
+        //- Return the list of fields required
+        virtual wordList fields() const;
 
[…]
