### 2024-11-12 b0fa890
Added ExaCA support to all tutorials (#39)
* Added ExaCA support to all tutorials with optional Allrun -withExaCA flag

Co-authored-by: Matt Rolchigo <65y@ornl.gov>

* cleaned up tutorials and added auto-enabled fucntion option with -withExaCA Allrun flag

* Update Allrun to remove exaca openmp setting

---------

Co-authored-by: Matt Rolchigo <65y@ornl.gov>
