### 2024-12-02 6f2fb7e
new utility to create scan paths (#42)
* new utility to create scan paths

* clean-up: removed nfs file clean
