+++ applications/utilities/createScanPath/createFields.H
@@ -1,0 +1,384 @@
+/*---------------------------------------------------------------------------*\
+  =========                 |
+  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
+   \\    /   O peration     | Website:  https://openfoam.org
+    \\  /    A nd           | Copyright (C) 2024 OpenFOAM Foundation
+     \\/     M anipulation  |
+-------------------------------------------------------------------------------
+                Copyright (C) 2023 Oak Ridge National Laboratory                
+-------------------------------------------------------------------------------
+License
+    This file is part of OpenFOAM.
+
+    OpenFOAM is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    OpenFOAM is distributed in the hope that it will be useful, but WITHOUT
+    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+    for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
+
+SourceFiles
+    createScanPath.C
+
+\*---------------------------------------------------------------------------*/
+
+#include "fvCFD.H"
+#include "OFstream.H"
+#include <algorithm>
+#include <fstream>
+#include <iomanip>
+#include <sstream>
+#include <vector>
+
+namespace Foam
+{
+
+/*---------------------------------------------------------------------------*\
+                            Struct Point Definition
+\*---------------------------------------------------------------------------*/
+
+struct Point
+{
+    scalar x;
+    scalar y;
+
+    // Default constructor
+    Point()
+        : x( 0.0 )
+        , y( 0.0 )
+    {
+    }
+
+    // Construct from x and y
+    Point( scalar xCoord, scalar yCoord )
+        : x( xCoord )
+        , y( yCoord )
+    {
+    }
+
+    // Function to rotate a point around a specified origin by a given angle
+    Point rotate( const Point& origin, scalar degrees ) const
+    {
+        scalar angle = degrees * ( M_PI / 180.0 );
+
+        scalar s = sin( angle );
+        scalar c = cos( angle );
+        scalar translatedX = x - origin.x;
+        scalar translatedY = y - origin.y;
+        scalar newX = translatedX * c - translatedY * s + origin.x;
+        scalar newY = translatedX * s + translatedY * c + origin.y;
+        return Point( newX, newY );
+    }
+
+    // Define a custom operator<< to print a Point
+    friend std::ostream& operator<<( std::ostream& os, const Point& point )
+    {
+        os << "(" << point.x << ", " << point.y << ")";
+        return os;
+    }
+};
+
+// Function to calculate the distance between two points
+scalar distance( const Point& p1, const Point& p2 )
+{
+    scalar dx = p1.x - p2.x;
+    scalar dy = p1.y - p2.y;
+    return std::sqrt( dx * dx + dy * dy );
+}
+
+
+/*---------------------------------------------------------------------------*\
+                            Struct Line Definition
+\*---------------------------------------------------------------------------*/
+
+struct Line
+{
+    Point start;
+    Point end;
+
+    // Default constructor
+    Line()
+        : start( Point() )
+        , end( Point() )
+    {
+    }
+
+    // Construct from two points
+    Line( Point startPoint, Point endPoint )
+        : start( startPoint )
+        , end( endPoint )
+    {
+    }
+
[…]
