### 2024-12-10 #7192 @srmnitc
@lukashof @FelixFroelich @sharanroongta, would any of you have time and be willing to review this submission for JOSS? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Thanks for your time!
