### 2025-02-10 #7192 @kincaidkc
@Rdfing would you have time and be willing to review this submission for JOSS? As you might know, they carry out their checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Thanks for your time!
