### 2025-02-11 #7192 @Rdfing
I am happy to review this paper. I however will be tied up for the coming two weeks.

> [@Rdfing](https://github.com/Rdfing) would you have time and be willing to review this submission for JOSS? As you might know, they carry out their checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Thanks for your time!
