### 2025-02-11 #7770 @editorialbot
[REVIEW]: AdditiveFOAM 1.0: A Continuum Physics Code for Additive Manufacturing
**Submitting author:** @colemanjs (<a href="http://orcid.org/0000-0002-7261-3143">John Coleman</a>)
**Repository:** https://github.com/ORNL/AdditiveFOAM.git
**Branch with paper.md** (empty if default branch): joss-2024
**Version:** 1.1.0
