### 2025-02-11 #7770 @srmnitc
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.1063/1.168744 is OK
> - 10.11578/dc.20211129.1 is OK
> - 10.1515/jnma-2023-0089 is OK
> - 10.1007/s11837-023-06279-3 is OK
> - 10.5281/zenodo.8034098 is OK
> - 10.1007/s40192-019-00164-1 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: FLOW-3D, Version 2023R1
> - No DOI given, and none found for title: Abaqus Documentation
> - No DOI given, and none found for title: ANSYS Documentation
> - No DOI given, and none found for title: Getting Started with Zoltan:  A Short Tutorial
>
> ❌ MISSING DOIs
>
> - 10.2172/1361589 may be a valid DOI for title: ALE3D: An arbitrary Lagrangian-Eulerian multi-phys...
> - 10.1016/j.commatsci.2022.111692 may be a valid DOI for title: ExaCA: A performance portable exascale cellular au...
>
> ❌ INVALID DOIs
>
> - https://doi.org/10.1016/j.addma.2024.104024 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.commatsci.2022.111904 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.addma.2023.103686 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.addma.2021.102578 is INVALID because of 'https://doi.org/' prefix
> ```

@colemanjs could you please fix the DOIs?
