### 2025-02-14 fff4c69
CI: run on ubuntu-latest
