### 2025-03-13 #7770 @Rdfing
@colemanjs could you please clarify the role and contribution of the coauthors of  Gerry L. Knapp, Benjamin Stump, and Alex Plotkowski? I did see them in the repository contributor.

@srmnitc I have completed my initial review. The package is in great shape. The repository and documentation meet the JOSS requirements, and the software appears robust and well-maintained.
