### 2025-03-14 #7770 @Rdfing
> Hi [@Rdfing](https://github.com/Rdfing), [@colemanjs](https://github.com/colemanjs) is out on leave for the next two weeks, so I can clarify. Most of the initial development for AdditiveFOAM was done on a separate private repository and that history was not included in the public release, so some commit history was lost. Here is a brief description for the roles of those coauthors:
>
> * G.L. Knapp: software testing and development, manuscript editing
> * B. Stump: software & algorithm development
> * A. Plotkowski: project management, initial software development, manuscript editing

Thanks for the clarification.
