### 2025-03-28 #7770 @sharanroongta
The documentation is well done! However, just as a suggestion maybe the installation instructions could include couple of statements (for users who aren't used to spack/docker).
All good from my side.
