### 2025-04-11 676cab2
Removed 'https://doi.org/' prefix from references
