+++ doc/paper.bib
@@ -86,7 +86,7 @@ volume = {81},
 pages = {104024},
 year = {2024},
 issn = {2214-8604},
-doi = {https://doi.org/10.1016/j.addma.2024.104024},
+doi = {10.1016/j.addma.2024.104024},
 url = {https://www.sciencedirect.com/science/article/pii/S2214860424000708},
 author = {Matt Rolchigo and John Coleman and Gerry L. Knapp and Alex Plotkowski}
 }
@@ -98,7 +98,7 @@ volume = {218},
 pages = {111904},
 year = {2023},
 issn = {0927-0256},
-doi = {https://doi.org/10.1016/j.commatsci.2022.111904},
+doi = {10.1016/j.commatsci.2022.111904},
 url = {https://www.sciencedirect.com/science/article/pii/S0927025622006152},
 author = {G.L. Knapp and J. Coleman and M. Rolchigo and M. Stoyanov and A. Plotkowski}
 }
@@ -110,7 +110,7 @@ author = {G.L. Knapp and J. Coleman and M. Rolchigo and M. Stoyanov and A. Plotk
   pages = {103686},
   year = {2023},
   issn = {2214-8604},
-  doi = {https://doi.org/10.1016/j.addma.2023.103686},
+  doi = {10.1016/j.addma.2023.103686},
   url = {https://www.sciencedirect.com/science/article/pii/S2214860423002993},
   author = {Michael P. Haines and Maxwell S. Moyle and Vitor V. Rielli and Vladimir Luzin and Nima Haghdadi and Sophie Primig}
 }
@@ -122,7 +122,7 @@ author = {G.L. Knapp and J. Coleman and M. Rolchigo and M. Stoyanov and A. Plotk
   pages = {102578},
   year = {2022},
   issn = {2214-8604},
-  doi = {https://doi.org/10.1016/j.addma.2021.102578},
+  doi = {10.1016/j.addma.2021.102578},
   url = {https://www.sciencedirect.com/science/article/pii/S2214860421007259},
   author = {MP Haines and F. List and K. Carver and DN Leonard and A. Plotkowski and CM Fancher and RR Dehoff and SS Babu}
 }
@@ -149,8 +149,7 @@ author = {G.L. Knapp and J. Coleman and M. Rolchigo and M. Stoyanov and A. Plotk
   year         = 2023,
   publisher    = {Zenodo},
   version      = {1.0.0},
-  doi          = {10.5281/zenodo.8034098},
-  url          = {https://doi.org/10.5281/zenodo.8034098}
+  doi          = {10.5281/zenodo.8034098}
 }
 
 @INPROCEEDINGS{zoltan,
