### 2025-04-11 #7770 @colemanjs
> The documentation is well done! However, just as a suggestion maybe the installation instructions could include couple of statements (for users who aren't used to spack/docker). All good from my side.

@sharanroongta - thank you for your review! Do you suggest listing the additional installation instructions directly in the README? We do link to the webpage for further installation instructions (https://ornl.github.io/AdditiveFOAM/docs/installation/#installation), but I realize that adds a layer of complexity for the user. Thanks!
