### 2025-04-30 b6791e8
Updates to AdditiveFOAM 1.1 as part of the JOSS review (#47)
* Update README.md (#35)

* Update README.md

Updated README.md with AdditiveFOAM logo

* Update README.md

updated website assets to ORNL-MDF project

* Update README.md order to prioritize documentation

* corrected explicit energy solve forced to remain forward euler. (#37)

* corrected explicit energy solve forced to remain forward euler. Also, added ExaCA support for all heat transfer tutorials

* made foamToExaCA a functionObject ExaCA to remain consistent with OpenFOAM library structure

* updated build instructions for CI

* 
[…]
