+++ applications/solvers/additiveFoam/foamToExaCA/foamToExaCA.C
@@ -1,445 +1,0 @@
-/*---------------------------------------------------------------------------*\
-  =========                 |
-  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
-   \\    /   O peration     | Website:  https://openfoam.org
-    \\  /    A nd           | Copyright (C) 2011-2022 OpenFOAM Foundation
-     \\/     M anipulation  |
--------------------------------------------------------------------------------
-                Copyright (C) 2023 Oak Ridge National Laboratory                
--------------------------------------------------------------------------------
-License
-    This file is part of OpenFOAM.
-
-    OpenFOAM is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    OpenFOAM is distributed in the hope that it will be useful, but WITHOUT
-    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
-    for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
-
-\*---------------------------------------------------------------------------*/
-
-#include "foamToExaCA/foamToExaCA.H"
-#include "addToRunTimeSelectionTable.H"
-#include "DynamicList.H"
-#include "interpolation.H"
-#include "labelVector.H"
-#include "pointMVCWeight.H"
-
-// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
-
-namespace Foam
-{
-    defineTypeNameAndDebug(foamToExaCA, 0);
-}
-
-// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
-
-Foam::foamToExaCA::foamToExaCA
-(
-    const volScalarField& T
-)
-:
-    IOdictionary
-    (
-        IOobject
-        (
-            "foamToExaCADict",
-            T.time().constant(),
-            T.mesh(),
-            IOobject::READ_IF_PRESENT,
-            IOobject::NO_WRITE
-        )
-    ),
-
-    mesh_(T.mesh()),
-
-    runTime_(T.time()),
-
-    T_(T),
-
-    vpi_(mesh_),
-
-    Tp_
-    (
-        IOobject
-        (
-            "Tp_",
-            runTime_.timeName(),
-            mesh_,
-            IOobject::NO_READ,
-            IOobject::NO_WRITE
-        ),
-        vpi_.interpolate(T_)
-    ),
-
-    execute_(false)
-{
-    if (this->headerOk())
-    {
-        execute_ = this->lookup<Switch>("execute");
-    }
-
-    initialize();
-}
-
-// * * * * * * * * * * * * * * Member Functions  * * * * * * * * * * * * * * //
-
-void Foam::foamToExaCA::initialize()
-{
-    if (!execute_)
-    {
-        return;
-    }
-
-    box_ = this->lookup("box");
-    iso_ = this->lookup<scalar>("isotherm");
-    dx_  = this->lookup<scalar>("dx");
-
-    treeBoundBox bb(box_);
-
-    // create a compact cell-stencil using the overlap sub-space
-    const pointField& points = mesh_.points();
-
-    treeBoundBox procBb(points);
-
-    const vector extend = 1e-10*vector::one;
-
-    if (procBb.overlaps(bb))
-    {
-        forAll(mesh_.cells(), celli)
-        {
-            treeBoundBox cellBb(point::max, point::min);
-
-            const labelList& vertices = mesh_.cellPoints()[celli];
-
[…]
