+++ .github/workflows/CI.yml
@@ -19,7 +19,7 @@ jobs:
         cxx: ['g++']
         cmake_build_type: ['Release']
        
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-latest
     container:
       image: docker.io/openfoam/openfoam10-paraview510
       options: --user root
@@ -29,11 +29,8 @@ jobs:
       - name: Build AdditiveFOAM
         run: |
           . /opt/openfoam10/etc/bashrc
-          cd applications/solvers/additiveFoam/movingHeatSource
-          wmake libso
-          cd -
           cd applications/solvers/additiveFoam
-          wmake
+          ./Allwmake
       - name: Test AdditiveFOAM
         run: |
           . /opt/openfoam10/etc/bashrc

[…]
