### 2025-04-30 4b7127d
Update paper.md
Added Matt Rolchigo and Samuel Reeve for tutorials and CI, respectively.
