### 2025-04-30 #7770 @colemanjs
@srmnitc – The capabilities described in the current JOSS article, including those developed during the review process, apply to AdditiveFOAM 1.1 rather than the previously released and archived AdditiveFOAM 1.0 on Zenodo. I’ve updated the paper title from “AdditiveFOAM 1.0: ...” to “AdditiveFOAM 1.1: ...”. Is there anything else I need to do to ensure this version update is properly reflected throughout the review process?

Additionally, I’d like to note that two contributors to the source code in this release were mistakenly omitted from the original author list. Matt Rolchigo developed the coupling to ExaCA for microstructure predictions, and Samuel Reeve implemented the testing framework and CI infrastructure for the repository. Is there a mechanism to formally add them as co-authors to the current JOSS article?
