### 2025-05-01 e83b904
Added Rolchigo to contributors list
