### 2025-05-01 7a009c2
added version and build info to additivefoam for tracking
