### 2025-05-02 #7770 @srmnitc
@colemanjs - for the first point. This is not a problem. Often new versions are released after the review. You could also keep just AdditiveFOAM on the title, if you wish. Otherwise, its ok this way.

The authors you suggested does seem to have contributions to the repo, so I think we should indeed add them. Please edit the paper to include the authors.
