### 2025-05-02 ee91ac7
removed SHA1 info and updated write until after setRootCase.H
