+++ applications/solvers/additiveFoam/additiveFoam.C
@@ -46,23 +46,18 @@ Description
 // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
 
 int main(int argc, char *argv[])
-{      
-    Info << nl << "--- AdditiveFOAM Information ---" << nl;
+{       
+    #include "postProcess.H"    
+    #include "setRootCase.H"
+    
+    Info << "AdditiveFOAM Information:" << nl;
     #ifdef ADDITIVEFOAM_VERSION
         Info << "Version:   " << ADDITIVEFOAM_VERSION << nl;
     #endif
     #ifdef ADDITIVEFOAM_GIT_DESCRIBE
-        Info << "Build:     " << ADDITIVEFOAM_GIT_DESCRIBE << nl;
+        Info << "Build:     " << ADDITIVEFOAM_GIT_DESCRIBE << nl << endl;
     #endif
-    #ifdef ADDITIVEFOAM_GIT_SHA1
-        Info << "Git SHA1:  " << ADDITIVEFOAM_GIT_SHA1 << nl << endl;
-    #endif
-
-    // * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-       
-    #include "postProcess.H"
-
-    #include "setRootCase.H"
+    
     #include "createTime.H"
     #include "createMesh.H"
     #include "createControl.H"
