### 2025-05-08 ecd2efa
corrected additivefoam static version and included git hash
