### 2025-05-08 cbf36d7
added version, build, and commit hash to AdditiveFOAM log files for version tracking (#50)
* added version and build info to additivefoam for tracking

* removed SHA1 info and updated write until after setRootCase.H

* corrected additivefoam static version and included git hash
