### 2025-05-08 cb831c5
Add joss paper to main for 1.1.0 release (#52)
---------

Co-authored-by: John Coleman <103219962+colemanjs@users.noreply.github.com>
Co-authored-by: Kellis Kincaid (khk) <kincaidkc@ornl.gov>
