+++ doc/paper.md
@@ -1,0 +1,75 @@
+---
+title: 'AdditiveFOAM: A Continuum Multiphysics Code for Additive Manufacturing'
+tags:
+  - Computational Fluid Dynamics
+  - Heat Transfer
+  - Mass Transfer
+  - Additive Manufacturing
+  - OpenFOAM
+authors:
+  - name: John Coleman
+    affiliation: "1"
+    corresponding: true
+  - name: Kellis Kincaid
+    affiliation: "2"
+  - name: Gerry L. Knapp
+    affiliation: "3"
+  - name: Benjamin Stump
+    affiliation: "1"
+  - name: Samuel Reeve
+    affiliation: "1"
+  - name: Matt Rolchigo
+    affiliation: "1"
+  - name: Alex Plotkowski
+    affiliation: "1"
+affiliations:
+ - name: Computational Science and Engineering Division, Oak Ridge National Laboratory
+   index: 1
+ - name: Nuclear Energy and Fuel Cycle Division, Oak Ridge National Laboratory
+   index: 2
+ - name: Material Science and Technology Division, Oak Ridge National Laboratory
+   index: 3
+date: 12 June 2024
+bibliography: paper.bib
+---
+
+# Summary
+
+AdditiveFOAM is a computational framework that simulates transport phenomena in Additive Manufacturing (AM) processes. It is built on OpenFOAM [@openfoam], the leading free, open-source software package for computational fluid dynamics (CFD). OpenFOAM offers an extensible platform for solving complex multiphysics problems using state-of-the-art finite volume methods. AdditiveFOAM leverages these capabilities to develop specialized tools aimed at addressing challenges in AM processing.
+
+Metal additive manufacturing, also known as metal 3D printing, is an advanced manufacturing technique that creates physical parts from a three-dimensional (3D) digital model by melting metal powder or wire feedstock. A significant area of research in metal AM focuses on process planning to mitigate anomalous features during printing that are deleterious to part performance (e.g., porosity and cracking), as well as controlling localized microstructure and material properties. Given the high costs and substantial time requirements associated with experimental methods for qualifying new materials and processes, there is a compelling incentive for researchers to utilize advanced computational simulations. In this context, AdditiveFOAM offers a simulation framework to better understand undesirable features in printing, thereby enhancing process planning and reducing the reliance on labor-intensive experimental campaigns.
+
+# Statement of need
+
+Numerical models for AM traditionally seek to represent physics at a single length scale depending on the target question. Two main categories of models with available software solutions are distinguished: 1) high-fidelity melt pool-scale models, and 2) part-scale heat transfer models.
+
+The first category contains models that seek to resolve all relevant physics within the melt pool. These models can produce excellent agreement with experiments and offer insights into the underlying causes of anomalous features during processing, but are generally expensive to run, limiting the scan length that can be simulated to a few millimeters with current computational resources. Due to their computational expense, these models are well-suited for use at specialized research institutions with large HPC infrastructure to answer target questions about melt pool scale physical phenomena. For example, ALE3D [@ALE3D] is a versatile multiphysics simulation tool that uses the Arbitrary Lagrangian-Eulerian approach and has been used for powder-resolved simulations of laser-material interactions in AM. However, ALE3D is a limited access code for use by United States Department of Defense/Energy laboratories and their contractors, while ALE3D4I is available for U.S. companies and academics through individual use agreements. Alternatively, FLOW-3D [@FLOW-3D] is a commercial CFD software known for its capabilities in simulating complex free-surface problems, and it has several specialized models for AM. However, FLOW-3D is proprietary, meaning its source code is not available for public inspection or modification, and users must purchase a license to use the software.
+
[…]
