+++ doc/paper.bib
@@ -1,0 +1,193 @@
+@article{openfoam,
+   abstract = {In this article the principles of the field operation and manipulation (FOAM) C++class library for continuum mechanics are outlined. Our intention is to make it as easy as possible to develop reliable and efficient computational continuum-mechanics codes: this is achieved by making the top-level syntax of the code as close as possible to conventional mathematical notation for tensors and partial differential equations. Object-orientation techniques enable the creation of data types that closely mimic those of continuum mechanics, and the operator overloading possible in C++ allows normal mathematical symbols to be used for the basic operations. As an example, the implementation of various types of turbulence modeling in a FOAM computational-fluid-dynamics code is discussed, and calculations performed on a standard test case, that of flow around a square prism, are presented. To demonstrate the flexibility of the FOAM library, codes for solving structures and magnetohydrodynamics are also presented with appropriate test case results given.},
+   author = {H. G. Weller and G. Tabor and H. Jasak and C. Fureby},
+   doi = {10.1063/1.168744},
+   issn = {08941866},
+   issue = {6},
+   journal = {Computers in Physics},
+   pages = {620},
+   title = {A tensorial approach to computational continuum mechanics using object-oriented techniques},
+   volume = {12},
+   year = {1998},
+}
+
+@techreport{ALE3D,
+  title={ALE3D: An arbitrary Lagrangian-Eulerian multi-physics code},
+  author={Noble, Charles R and Anderson, Andrew T and Barton, Nathan R and Bramwell, Jamie A and Capps, Arlie and Chang, Michael H and Chou, Jin J and Dawson, David M and Diana, Emily R and Dunn, Timothy A and others},
+  year={2017},
+  institution={Lawrence Livermore National Lab.(LLNL), Livermore, CA (United States)}
+}
+
+@Manual{FLOW-3D,
+  author = {Flow Science, Inc.},
+  title = {FLOW-3D, Version~2023R1},
+  year = {2023},
+  address = {Santa Fe, NM},
+  url = {https://www.flow3d.com/}
+}
+
+@manual{abaqus,
+  title        = {Abaqus Documentation},
+  author       = {{Dassault Systèmes Simulia Corp.}},
+  organization = {Dassault Systèmes},
+  address      = {Providence, RI, USA},
+  year         = {2024},
+  note         = {Version 2024},
+  url          = {https://www.3ds.com/products-services/simulia/products/abaqus/}
+}
+
+@manual{ansys,
+  title        = {ANSYS Documentation},
+  author       = {{ANSYS Inc.}},
+  organization = {ANSYS Inc.},
+  address      = {Canonsburg, PA, USA},
+  year         = {2024},
+  note         = {Release 2024 R1},
+  url          = {https://www.ansys.com/products/structures/ansys-mechanical}
+}
+
+@software{adamantine,
+  author = {Turcksin, Bruno and DeWitt, Stephen},
+  doi = {10.11578/dc.20211129.1},
+  title = {{adamantine}},
+  url = {https://github.com/adamantine-sim/adamantine}
+}
+
+@Article{deal.ii,
+  title   = {The \texttt{deal.II} Library, Version 9.5},
+  author  = {Daniel Arndt and Wolfgang Bangerth and Maximilian Bergbauer and
+             Marco Feder and Marc Fehling and Johannes Heinz and
+             Timo Heister and Luca Heltai and Martin Kronbichler and
[…]
