### 2025-05-12 3698451
added build flags to enable optional git information from Allwmake, and added AdditivefoamInfo class to write information to log
