### 2025-05-12 64a3cd0
Build info fixup (#54)
* added build flags to enable optional git information from Allwmake, and added AdditivefoamInfo class to write information to log

* fixup: removed .nfs file

* fixup: added undefined to git info in Allwmake for shell failure
