### 2025-05-14 6ab1e18
Revert "Add joss paper to main for 1.1.0 release (#52)"
This reverts commit cb831c5be60ecaff5fedf3c6061587fad551159b.
