### 2025-05-14 0d389e2
Revert "Add joss paper to main for 1.1.0 release (#52)" (#55)
This reverts commit cb831c5be60ecaff5fedf3c6061587fad551159b.
