### 2025-05-14 9a2a592
fixup: use meta DOI for readme
