### 2025-05-14 13761ec
Add changelog for 1.1
