### 2025-05-19 ORNL/AdditiveFOAM#59 @srmnitc
Changes to the JOSS paper
Part of review over at https://github.com/openjournals/joss-reviews/issues/7770

Some small recommendations for the paper:
- maybe line 36 could be strengthened with a citation?
- line 52, citation Turcksin & DeWitt is not correct, see also line 164
- line 54, please recheck. '..including the deal.II finite...'
- line 57, needs citation
- line 60, probably also needs a citation
- line 77-78, the new terms might be better with references
- line 106 - link could be embedded into the text
- references: journal abbreviations and full names are both used. Please take a look and choose one.
