### 2025-05-19 ff2773c
corrected adamantine citation to JOSS article
