+++ doc/paper.bib
@@ -46,11 +46,17 @@
   url          = {https://www.ansys.com/products/structures/ansys-mechanical}
 }
 
-@software{adamantine,
-  author = {Turcksin, Bruno and DeWitt, Stephen},
-  doi = {10.11578/dc.20211129.1},
-  title = {{adamantine}},
-  url = {https://github.com/adamantine-sim/adamantine}
+@article{Turcksin2024,
+  author    = {Bruno Turcksin and Stephen DeWitt},
+  title     = {Adamantine 1.0: A Thermomechanical Simulator for Additive Manufacturing},
+  journal   = {Journal of Open Source Software},
+  volume    = {9},
+  number    = {102},
+  pages     = {7017},
+  year      = {2024},
+  publisher = {The Open Journal},
+  doi       = {10.21105/joss.07017},
+  url       = {https://doi.org/10.21105/joss.07017},
 }
 
 @Article{deal.ii,
