+++ doc/paper.md
@@ -69,7 +69,7 @@ AdditiveFOAM was designed to be used openly by researchers, industry, and academ
 
 - Plotkowski et al. [@plotkowski] used AdditiveFOAM to confirm experimentally observed grain growth in Fe-Si steel during the Laser Engineered Net Shaping (LENS) process. 
 
-The free, open-source nature of AdditiveFOAM will continue to enable scientific explorations in AM processing. Documentation for AdditiveFOAM is hosted on GitHub pages with the link: https://ornl.github.io/AdditiveFOAM/. 
+The free, open-source nature of AdditiveFOAM will continue to enable scientific explorations in AM processing. Documentation for AdditiveFOAM is hosted on [GitHub Pages](https://ornl.github.io/AdditiveFOAM/).
 
 # Acknowledgements
 
