### 2025-05-19 2da3693
fixup: adamantine description
