+++ doc/paper.md
@@ -52,7 +52,7 @@ Numerical models for AM traditionally seek to represent physics at a single leng
 
 The first category contains models that seek to resolve all relevant physics within the melt pool. These models can produce excellent agreement with experiments and offer insights into the underlying causes of anomalous features during processing, but are generally expensive to run, limiting the scan length that can be simulated to a few millimeters with current computational resources [@Yan2020]. Due to their computational expense, these models are well-suited for use at specialized research institutions with large HPC infrastructure to answer target questions about melt pool scale physical phenomena. For example, ALE3D [@ALE3D] is a versatile multiphysics simulation tool that uses the Arbitrary Lagrangian-Eulerian approach and has been used for powder-resolved simulations of laser-material interactions in AM. However, ALE3D is a limited access code for use by United States Department of Defense/Energy laboratories and their contractors, while ALE3D4I is available for U.S. companies and academics through individual use agreements. Alternatively, FLOW-3D [@FLOW-3D] is a commercial CFD software known for its capabilities in simulating complex free-surface problems, and it has several specialized models for AM. However, FLOW-3D is proprietary, meaning its source code is not available for public inspection or modification, and users must purchase a license to use the software.
 
-The second category contains models that simplify melt pool physics to rapidly simulate heat transport, residual stress, and distortion across an entire part. Commercial finite element thermomechanics software solutions built on Abaqus [@abaqus] and Ansys [@ansys] are available; however, these software packages are not open and free to use and develop upon. Alternatively, Adamantine [@Turcksin2024] is a thermomechanics simulation tool built on an open-source software stack designed for high-performance computing across various architectures, including the deal.II finite element software package [@deal.ii], providing an alternative to proprietary software solutions.
+The second category contains models that simplify melt pool physics to rapidly simulate heat transport, residual stress, and distortion across an entire part. Commercial finite element thermomechanics software solutions built on Abaqus [@abaqus] and Ansys [@ansys] are available; however, these software packages are not open and free to use and develop upon. Alternatively, Adamantine [@Turcksin2024] is a thermomechanics simulation tool built on the open-source deal.II finite element software package [@deal.ii] designed for high-performance computing across architectures, providing an alternative to proprietary software solutions.
 
 There is an established need for intermediate frameworks that accurately predict melt pool shape, thermal gradients, and other meso-scale quantities across an entire part [@DEBROY2018112]. These models can inform process design decisions through heuristic estimations of anomalous printing features (e.g., keyhole formation and lack-of-fusion) and predict the final microstructure and material properties of AM components [@WEI2021100703]. AdditiveFOAM addresses these challenges, featuring a volumetric source term in the energy equation for multiple heat sources, optional Marangoni-driven fluid flow, and a scheme for implementing tabulated thermodynamic pathways for metal alloys.
 
