### 2025-05-19 00db238
removed abbreviations from citations
