+++ doc/paper.bib
@@ -137,7 +137,7 @@ author = {G.L. Knapp and J. Coleman and M. Rolchigo and M. Stoyanov and A. Plotk
   author = {Plotkowski, A. and Coleman, J. and Fancher, C. M. and Haines, M. P. and Whetten, S. R. and Kustas, A. B.},
   year = {2024},
   title = {Grain Structure Evolution in Fe-6Si During Directed Energy Deposition},
-  journal = {JOM},
+  journal = {Journal of The Minerals, Metals \& Materials Society},
   pages = {1031},
   volume = {76},
   number = {3},
@@ -159,13 +159,13 @@ author = {G.L. Knapp and J. Coleman and M. Rolchigo and M. Stoyanov and A. Plotk
 }
 
 @INPROCEEDINGS{zoltan,
- AUTHOR = "K.D. Devine and E.G. Boman and L.A. Riesen and U.V. Catalyurek 
+ author = "K.D. Devine and E.G. Boman and L.A. Riesen and U.V. Catalyurek 
           and C. Chevalier",
- TITLE = "Getting Started with Zoltan:  A Short Tutorial",
- BOOKTITLE = "Proc. of 2009 Dagstuhl Seminar on Combinatorial Scientific 
-              Computing", 
- YEAR = "2009",
- NOTE = "Also available as Sandia National Labs Tech Report SAND2009-0578C."
+ title = "Getting Started with Zoltan:  A Short Tutorial",
+ booktitle = {Proceedings of the 2009 Dagstuhl Seminar on Combinatorial Scientific Computing},
+
+ year = "2009",
+ note = "Also available as Sandia National Labs Tech Report SAND2009-0578C."
 }
 
 @article{amb2018,
