### 2025-05-19 ec56c30
added supporting references for modeling costs, needs, and usage
