+++ doc/paper.bib
@@ -11,6 +11,17 @@
    year = {1998},
 }
 
+@article{Yan2020,
+  author    = {Yan, Wentao and Lu, Yan and Jones, Kevontrez and Yang, Zhuo and Fox, Jason C. and Witherell, Paul W. and Wagner, Gregory J. and Liu, Wing Kam},
+  title     = {Data-driven characterization of computational models for powder-bed-fusion additive manufacturing},
+  journal   = {Additive Manufacturing},
+  year      = {2020},
+  volume    = {36},
+  pages     = {101503},
+  doi       = {10.1016/j.addma.2020.101503},
+  url       = {https://doi.org/10.1016/j.addma.2020.101503}
+}
+
 @techreport{ALE3D,
   title={ALE3D: An arbitrary Lagrangian-Eulerian multi-physics code},
   author={Noble, Charles R and Anderson, Andrew T and Barton, Nathan R and Bramwell, Jamie A and Capps, Arlie and Chang, Michael H and Chou, Jin J and Dawson, David M and Diana, Emily R and Dunn, Timothy A and others},
@@ -73,6 +84,30 @@
   volume  = {31},
   number  = {3},
   url     = {https://dealii.org/deal95-preprint.pdf}
+}
+
+@article{DEBROY2018112,
+  author  = {T. DebRoy and H.~L. Wei and J.~S. Zuback and T. Mukherjee and J.~W. Elmer and J.~O. Milewski and A.~M. Beese and A. Wilson-Heid and A. De and W. Zhang},
+  title   = {Additive manufacturing of metallic components – Process, structure and properties},
+  journal = {Progress in Materials Science},
+  volume  = {92},
+  pages   = {112--224},
+  year    = {2018},
+  issn    = {0079-6425},
+  doi     = {10.1016/j.pmatsci.2017.10.001},
+  url     = {https://www.sciencedirect.com/science/article/pii/S0079642517301172},
+}
+
+@article{WEI2021100703,
+  author  = {H.L. Wei and T. Mukherjee and W. Zhang and J.S. Zuback and G.L. Knapp and A. De and T. DebRoy},
+  title   = {Mechanistic models for additive manufacturing of metallic components},
+  journal = {Progress in Materials Science},
+  volume  = {116},
+  pages   = {100703},
+  year    = {2021},
+  issn    = {0079-6425},
+  doi     = {10.1016/j.pmatsci.2020.100703},
+  url     = {https://www.sciencedirect.com/science/article/pii/S0079642520300670},
 }
 
 @article{exaca,
