### 2025-05-19 ORNL/AdditiveFOAM#59 @colemanjs
1. Lines 36, 57, and 60 have been strengthened with citations [ec56c303f8b156d2bcabc6b19b5ce200a489163d]
2. Abbreviation has been removed from citations [00db2388f8d74966144752a40b17d01578863979]
3. Line 106 has been updated to embed the github pages link [4d1a258b2b10447eefe11dc55a0ea7fa16efcd3d]
4. The adamantine citation has been fixed (Line 57) [ff2773c052cc132e6a150ceb665eb5c0dee192fd]

For lines 77-78: I am not sure we need references to the specific processing methods, they are common methods in AM
