### 2025-05-19 #7770 @colemanjs
> [@colemanjs](https://github.com/colemanjs) [@streeve](https://github.com/streeve) I made a small issue with some changes to the paper, could you please take a look?

Thank you for the review. I believe that we have addressed all changed in the current draft.
