### 2025-05-19 ORNL/AdditiveFOAM#59 @streeve
I went ahead and addressed Line 54 for clarity as well
