### 2025-05-20 #7770 @srmnitc
@colemanjs and @streeve editorialbot seems to be correct:

- 10.2172/1361589 may be a valid DOI for title: ALE3D: An arbitrary Lagrangian-Eulerian multi-phys...
- 10.1016/j.commatsci.2022.111692 may be a valid DOI for title: ExaCA: A performance portable exascale cellular au...

both seem to be valid DOIs which are not in the paper. Could you please take a look?
