+++ doc/paper.bib
@@ -23,11 +23,14 @@
 }
 
 @techreport{ALE3D,
-  title={ALE3D: An arbitrary Lagrangian-Eulerian multi-physics code},
-  author={Noble, Charles R and Anderson, Andrew T and Barton, Nathan R and Bramwell, Jamie A and Capps, Arlie and Chang, Michael H and Chou, Jin J and Dawson, David M and Diana, Emily R and Dunn, Timothy A and others},
-  year={2017},
-  institution={Lawrence Livermore National Lab.(LLNL), Livermore, CA (United States)}
-}
+  author       = {Noble, Charles R. and Anderson, Andrew T. and Barton, Nathan R. and Bramwell, Jamie A. and Capps, Arlie and Chang, Michael H. and Chou, Jin J. and Dawson, David M. and Diana, Emily R. and Dunn, Timothy A. and others},
+  title        = {ALE3D: An Arbitrary Lagrangian-Eulerian Multi-Physics Code},
+  institution  = {Lawrence Livermore National Lab. (LLNL), Livermore, CA (United States)},
+  doi          = {10.2172/1361589},
+  url          = {https://www.osti.gov/biblio/1361589},
+  place        = {United States},
+  year         = {2017},
+  month        = {05}}
 
 @Manual{FLOW-3D,
   author = {Flow Science, Inc.},
@@ -117,7 +120,8 @@
   volume={214},
   pages={111692},
   year={2022},
-  publisher={Elsevier}
+  issn = {0927-0256},
+  doi={10.1016/j.commatsci.2022.111692}
 }
 
 @article{texture-regimes,
