### 2025-05-20 #7770 @colemanjs
@srmnitc - thank you for your help throughout this process. I have a small correction for the final proof (the affiliation "1" had a typo that was corrected in the most recent PR to the paper branch). Can we generate a new final proof with the updated change?
